fix(auth): stop the autoCleanup registry test from leaking Clerk users [GH-000]
The whole-branch review found autoCleanup.spec.ts running inside the real
Playwright suite, sharing the same worker process and module-level registry
as live specs. Its toHaveLength(2) assertion failed against the real
(larger) registry, and its stub-deleter drain call spliced out -- and
thereby lost -- real entries instead of deleting them, leaking Clerk users
into a capped shared instance.

- Move all three registry cases into apps/auth/tests/autoCleanup.test.ts
  (vitest), resetting the singleton registry between cases via
  beforeEach/afterEach so they can't leak into each other.
- Delete the Playwright spec; keep fixtures/autoCleanup.ts (the fixture
  itself is correct and is what real specs import).
- Fix five specs that called createTestUser but imported the stock `test`
  from @playwright/test instead of the autoCleanup fixture -- coverage
  today was accidental (workers: 1 + file co-residency), not declared.
- Compute STUDIO_CHANGED in scripts/detect-changes.sh so a studio-only
  change is no longer misclassified as docs-only by the pre-push hook.
- Correct two present-tense doc references to the deleted docker/watcher.mjs
  to describe the compose.yml healthcheck that replaced it.
- Correct two comments that overstated behavior: the CI sweeper's
  --older-than-hours window (documents the invariant holds one run late)
  and the admin env page's auth claim (the guard is client-side, not
  server-side).
