+++ apps/auth/e2e/permission-management.spec.ts
@@ -1,6 +1,7 @@
 import path from "node:path";
 
-import { expect, test, type BrowserContext, type Page } from "@playwright/test";
+import type { BrowserContext, Page } from "@playwright/test";
+import { expect, test } from "./fixtures/autoCleanup";
 
 import { cleanupTestData } from "./helpers/cleanup";
 import {
