+++ apps/auth/e2e/receipt-reference-flow.spec.ts
@@ -2,7 +2,7 @@ import { createHash } from "node:crypto";
 import { readFileSync } from "node:fs";
 import path from "node:path";
 
-import { expect, test } from "@playwright/test";
+import { expect, test } from "./fixtures/autoCleanup";
 
 import { cleanupTestData } from "./helpers/cleanup";
 import {
