+++ scripts/detect-changes.sh
@@ -24,6 +24,7 @@ is_changed() {
 
 DEPLOY_CHANGED=false
 STORE_CHANGED=false
+STUDIO_CHANGED=false
 LANDING_CHANGED=false
 PAYMENTS_CHANGED=false
 ADMIN_CHANGED=false
@@ -34,6 +35,7 @@ CODE_CHANGED=false
 
 is_changed '^(Dockerfile|docker/|\.dockerignore)' && DEPLOY_CHANGED=true
 is_changed '^apps/store/' && STORE_CHANGED=true
+is_changed '^apps/studio/' && STUDIO_CHANGED=true
 is_changed '^apps/landing/' && LANDING_CHANGED=true
 is_changed '^apps/payments/' && PAYMENTS_CHANGED=true
 is_changed '^apps/admin/' && ADMIN_CHANGED=true
@@ -43,7 +45,8 @@ is_changed '^(package\.json|pnpm-lock\.yaml|tsconfig.*\.json|eslint\.config\..*|
 is_changed '\.(ts|tsx|js|jsx)$' && CODE_CHANGED=true
 
 DOCS_ONLY=true
-if [ "$DEPLOY_CHANGED" = true ] || [ "$STORE_CHANGED" = true ] || [ "$LANDING_CHANGED" = true ] || \
+if [ "$DEPLOY_CHANGED" = true ] || [ "$STORE_CHANGED" = true ] || [ "$STUDIO_CHANGED" = true ] || \
+   [ "$LANDING_CHANGED" = true ] || \
    [ "$PAYMENTS_CHANGED" = true ] || [ "$ADMIN_CHANGED" = true ] || \
    [ "$AUTH_CHANGED" = true ] || \
    [ "$PACKAGES_CHANGED" = true ] || [ "$TOOLING_CHANGED" = true ]; then
@@ -52,6 +55,7 @@ fi
 
 echo "DEPLOY_CHANGED=$DEPLOY_CHANGED"
 echo "STORE_CHANGED=$STORE_CHANGED"
+echo "STUDIO_CHANGED=$STUDIO_CHANGED"
 echo "LANDING_CHANGED=$LANDING_CHANGED"
 echo "PAYMENTS_CHANGED=$PAYMENTS_CHANGED"
 echo "ADMIN_CHANGED=$ADMIN_CHANGED"
